refactor: add AbstractIcon class and make Icon extend it
